dom/serviceworkers/test/test_not_intercept_plugin.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<!DOCTYPE HTML>
<html>
<head>
  <title>Bug 1187766 - Test loading plugins scenarios with fetch interception.</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none"></div>
<pre id="test"></pre>
<script src="utils.js"></script>
<script class="testbody" type="text/javascript">
  SimpleTest.requestCompleteLog();

  var registration;
  function simpleRegister() {
    var p = navigator.serviceWorker.register("./fetch/plugin/worker.js", { scope: "./fetch/plugin/" });
    return p.then(function(swr) {
      registration = swr;
      return waitForState(swr.installing, 'activated');
    });
  }

  function unregister() {
    return registration.unregister().then(function(success) {
      ok(success, "Service worker should be unregistered successfully");
    }, function(e) {
      dump("SW unregistration error: " + e + "\n");
    });
  }

  function testPlugins() {
    var p = new Promise(function(resolve, reject) {
      window.onmessage = function(e) {
        if (e.data.status == "ok") {
          ok(e.data.result, e.data.message);
        } else if (e.data.status == "done") {
          window.onmessage = null;
          w.close();
          resolve();
        }
      }
    });

    var w = window.open("fetch/plugin/plugins.html");
    return p;
  }

  function runTest() {
    simpleRegister()
      .then(testPlugins)
      .then(unregister)
      .then(function() {
        SimpleTest.finish();
      }).catch(function(e) {
        ok(false, "Some test failed with error " + e);
        SimpleTest.finish();
      });
  }

  SimpleTest.waitForExplicitFinish();
  SpecialPowers.pushPrefEnv({"set": [
    ["dom.serviceWorkers.exemptFromPerDomainMax", true],
    ["dom.serviceWorkers.enabled", true],
    ["dom.serviceWorkers.testing.enabled", true],
  ]}, runTest);
</script>
</pre>
</body>
</html>
